refine the test page and move discalimer
